requirements.txt
pyttsx3==2.90
SpeechRecognition==3.8.1
pywhatkit==5.4
yfinance==0.2.26
pyjokes==0.6.0
wikipedia-api==0.5.4
requests==2.31.0
pyaudio==0.2.14
setuptools==75.1.0
